+++ website/src/logos/index.ts
@@ -1,4 +1,5 @@
 export * from './AlgoliaLogo';
+export * from './AnthropicLogo';
 export * from './BoltLogo';
 export * from './BuilderLogo';
 export * from './DailyDevLogo';
@@ -7,6 +8,7 @@ export * from './HdmLogo';
 export * from './LambdaTestLogo';
 export * from './MotionLogo';
 export * from './NetlifyLogo';
+export * from './OpenAILogo';
 export * from './PaceLogo';
 export * from './StainlessLogo';
 export * from './TestMuAiLogo';